Create template for video pages
